feat: modal de criação avançada, pesos em float e otimização do UI
- Implementação da criação de módulos com auto-replicação de parâmetros para as 4 fases práticas.

- Alteração da propriedade Weight para float nos modelos e DTOs, com respetiva migração na base de dados.

- Correção da responsividade e compactação visual das datas no construtor.

+++ Gamp-SCPeriop.Shared/Data/Module.cs
@@ -11,7 +11,7 @@ namespace GamP_SCPeriop.Shared.Data
         public List<ModuleComponent> Components { get; set; } = new();
         public int? PathwayId { get; set; }
         public List<ModuleStageTimeline>? StageTimelines { get; set; }
-        public int Weight { get; set; } = 1;
+        public float Weight { get; set; } = 1;
         public bool IsFromTemplate { get; set; }
 
         [NotMapped]
