+++ Gamp-SCPeriop.Shared/Data/Template/ComponentTemplate.cs
@@ -13,7 +13,7 @@ namespace GamP_SCPeriop.Shared.Data.Template
         // A fase a que pertence (Observação, Prática, etc.)
         public ModuleStage Stage { get; set; }
 
-        public int Weight { get; set; } = 1; // Para o cálculo de percentagens do teu print
+        public float Weight { get; set; } = 1; // Para o cálculo de percentagens do teu print
 
         public string? PdfFilePath { get; set; }
