+++ Gamp-SCPeriop.Shared/Entity/Model/BuilderComponentDto.cs
@@ -11,7 +11,7 @@ namespace GamP_SCPeriop.Shared.Entity.Model
         public string Title { get; set; } = string.Empty;
         public string? Description { get; set; }
         public ModuleStage Stage { get; set; }
-        public int Weight { get; set; } = 1;
+        public float Weight { get; set; } = 1;
         public int? ParentComponentId { get; set; }
         public string? PdfFilePath { get; set; }
         public bool IsFromTemplate { get; set; }
